requirements.txt
langgraph
langchain
langchain-openai
streamlit
pytesseract
Pillow
opencv-python
python-dotenv